Remove separate deprecated files.
