+++ coil-compose-base/api/coil-compose-base.api
@@ -12,6 +12,31 @@ public final class coil/compose/AsyncImagePainter : androidx/compose/ui/graphics
 	public fun onRemembered ()V
 }
 
+public abstract interface class coil/compose/AsyncImagePainter$ExecuteCallback {
+	public static final field Companion Lcoil/compose/AsyncImagePainter$ExecuteCallback$Companion;
+	public static final field Default Lcoil/compose/AsyncImagePainter$ExecuteCallback;
+	public abstract fun invoke (Lcoil/compose/AsyncImagePainter$Snapshot;Lcoil/compose/AsyncImagePainter$Snapshot;)Z
+}
+
+public final class coil/compose/AsyncImagePainter$ExecuteCallback$Companion {
+}
+
+public final class coil/compose/AsyncImagePainter$Snapshot {
+	public static final field $stable I
+	public synthetic fun <init> (Lcoil/compose/AsyncImagePainter$State;Lcoil/request/ImageRequest;JLkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public final fun component1 ()Lcoil/compose/AsyncImagePainter$State;
+	public final fun component2 ()Lcoil/request/ImageRequest;
+	public final fun component3-NH-jbRc ()J
+	public final fun copy-cSwnlzA (Lcoil/compose/AsyncImagePainter$State;Lcoil/request/ImageRequest;J)Lcoil/compose/AsyncImagePainter$Snapshot;
+	public static synthetic fun copy-cSwnlzA$default (Lcoil/compose/AsyncImagePainter$Snapshot;Lcoil/compose/AsyncImagePainter$State;Lcoil/request/ImageRequest;JILjava/lang/Object;)Lcoil/compose/AsyncImagePainter$Snapshot;
+	public fun equals (Ljava/lang/Object;)Z
+	public final fun getRequest ()Lcoil/request/ImageRequest;
+	public final fun getSize-NH-jbRc ()J
+	public final fun getState ()Lcoil/compose/AsyncImagePainter$State;
+	public fun hashCode ()I
+	public fun toString ()Ljava/lang/String;
+}
+
 public abstract class coil/compose/AsyncImagePainter$State {
 	public static final field $stable I
 	public abstract fun getPainter ()Landroidx/compose/ui/graphics/painter/Painter;
@@ -59,10 +84,7 @@ public final class coil/compose/AsyncImagePainter$State$Success : coil/compose/A
 
 public final class coil/compose/AsyncImagePainterKt {
 	public static final fun rememberAsyncImagePainter (Ljava/lang/Object;Lcoil/ImageLoader;Landroidx/compose/runtime/Composer;I)Lcoil/compose/AsyncImagePainter;
-}
-
-public final class coil/compose/DeprecatedKt {
-	public static final fun rememberImagePainter (Lcoil/request/ImageRequest;Lcoil/ImageLoader;Landroidx/compose/runtime/Composer;I)Lcoil/compose/AsyncImagePainter;
-	public static final fun rememberImagePainter (Ljava/lang/Object;Lcoil/ImageLoader;Lkotlin/jvm/functions/Function1;Landroidx/compose/runtime/Composer;II)Lcoil/compose/AsyncImagePainter;
+	public static final fun rememberImagePainter (Lcoil/request/ImageRequest;Lcoil/ImageLoader;Lcoil/compose/AsyncImagePainter$ExecuteCallback;Landroidx/compose/runtime/Composer;II)Lcoil/compose/AsyncImagePainter;
+	public static final fun rememberImagePainter (Ljava/lang/Object;Lcoil/ImageLoader;Lcoil/compose/AsyncImagePainter$ExecuteCallback;Lkotlin/jvm/functions/Function1;Landroidx/compose/runtime/Composer;II)Lcoil/compose/AsyncImagePainter;
 }
