+++ coil-compose-base/src/main/java/coil/compose/AsyncImagePainter.kt
@@ -1,3 +1,5 @@
+@file:Suppress("DEPRECATION_ERROR", "unused", "UNUSED_PARAMETER")
+
 package coil.compose
 
 import android.graphics.drawable.BitmapDrawable
@@ -27,6 +29,7 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalInspectionMode
 import coil.ImageLoader
+import coil.compose.AsyncImagePainter.ExecuteCallback
 import coil.compose.AsyncImagePainter.State
 import coil.decode.DataSource
 import coil.request.ErrorResult
@@ -190,6 +193,33 @@ class AsyncImagePainter internal constructor(
         }
     }
 
+    @Deprecated(
+        message = "ExecuteCallback no longer has any effect. To skip waiting for size resolution " +
+            "use ImageRequest.Builder.size(OriginalSize).",
+        level = DeprecationLevel.ERROR
+    )
+    fun interface ExecuteCallback {
+
+        operator fun invoke(previous: Snapshot?, current: Snapshot): Boolean
+
+        companion object {
+            @JvmField val Default = ExecuteCallback { previous, current ->
+                previous?.request != current.request
+            }
+        }
+    }
+
+    @Deprecated(
+        message = "ExecuteCallback no longer has any effect. To skip waiting for size resolution " +
+            "use ImageRequest.Builder.size(OriginalSize).",
+        level = DeprecationLevel.ERROR
+    )
+    data class Snapshot(
+        val state: State,
+        val request: ImageRequest,
+        val size: Size,
+    )
+
     /**
      * The current state of the [AsyncImagePainter].
      */
@@ -311,3 +341,52 @@ internal fun requestOf(model: Any?): ImageRequest {
         ImageRequest.Builder(LocalContext.current).data(model).build()
     }
 }
+
+@Deprecated(
+    message = "ImagePainter has been renamed to AsyncImagePainter.",
+    replaceWith = ReplaceWith(
+        expression = "AsyncImagePainter",
+        imports = ["coil.compose.AsyncImagePainter"]
+    ),
+    level = DeprecationLevel.ERROR
+)
+typealias ImagePainter = AsyncImagePainter
+
+@Deprecated(
+    message = "ImagePainter has been renamed to AsyncImagePainter.",
+    replaceWith = ReplaceWith(
+        expression = "rememberAsyncImagePainter(" +
+            "ImageRequest.Builder(LocalContext.current).data(data).apply(builder).build(), " +
+            "imageLoader)",
+        imports = ["coil.compose.rememberAsyncImagePainter", "coil.request.ImageRequest"]
+    ),
+    level = DeprecationLevel.ERROR
+)
+@Composable
+inline fun rememberImagePainter(
+    data: Any?,
+    imageLoader: ImageLoader,
+    onExecute: ExecuteCallback = ExecuteCallback.Default,
+    builder: ImageRequest.Builder.() -> Unit = {},
+) = rememberAsyncImagePainter(
+    model = ImageRequest.Builder(LocalContext.current).data(data).apply(builder).build(),
+    imageLoader = imageLoader
+)
+
+@Deprecated(
+    message = "ImagePainter has been renamed to AsyncImagePainter.",
+    replaceWith = ReplaceWith(
+        expression = "rememberAsyncImagePainter(request, imageLoader)",
+        imports = ["coil.compose.rememberAsyncImagePainter"]
+    ),
+    level = DeprecationLevel.ERROR
+)
+@Composable
+fun rememberImagePainter(
+    request: ImageRequest,
+    imageLoader: ImageLoader,
+    onExecute: ExecuteCallback = ExecuteCallback.Default
+) = rememberAsyncImagePainter(
+    model = request,
+    imageLoader = imageLoader
+)
